// src/components/Models/ComparisonOperator.ts
export enum ComparisonOperator {
	Equal = 'Equal',
	GreaterThan = 'GreaterThan',
	LessThan = 'LessThan',
	LessThanOrEqual = 'LessThanOrEqual',
	GreaterThanOrEqual = 'GreaterThanOrEqual',
	NotEqual = 'NotEqual',
	IsNull = 'IsNull',
	IsNotNull = 'IsNotNull',
	Like = 'Like',
	NotLike = 'NotLike',
	In = 'In',
	NotIn = 'NotIn',
	Between = 'Between',
	NotBetween = 'NotBetween',
}

export default ComparisonOperator;