// src/components/Models/DataType.ts
export enum DataType {
	TinyInt = 'TinyInt',
	TinyIntUnsigned = 'TinyIntUnsigned',
	SmallInt = 'SmallInt',
	SmallIntUnsigned = 'SmallIntUnsigned',
	MediumInt = 'MediumInt',
	MediumIntUnsigned = 'MediumIntUnsigned',
	Int = 'Int',
	IntUnsigned = 'IntUnsigned',
	BigInt = 'BigInt',
	BigIntUnsigned = 'BigIntUnsigned',
	Numeric = 'Numeric',
	Single = 'Single',
	Double = 'Double',
	Percent = 'Percent',
	Phone = 'Phone',
	Money = 'Money',
	SmallMoney = 'SmallMoney',
	Binary = 'Binary',
	Date = 'Date',
	DateTime = 'DateTime',
	Time = 'Time',
	TimeUtc = 'TimeUtc',
	DateTimeUtc = 'DateTimeUtc',
	SmallYear = 'SmallYear',
	Year = 'Year',
	TimeSpan = 'TimeSpan',
	Boolean = 'Boolean',
	MacAddress = 'MacAddress',
	IpAddress = 'IpAddress',
	Email = 'Email',
	Uri = 'Uri',
	HostName = 'HostName',
	Bit = 'Bit',
	VaryingBit = 'VaryingBit',
	UUID = 'UUID',
	Json = 'Json',
	Xml = 'Xml',
	String = 'String',
	AnsiString = 'AnsiString',
	FixedString = 'FixedString',
	FixedAnsiString = 'FixedAnsiString',
	VaryingString = 'VaryingString',
	VaryingAnsiString = 'VaryingAnsiString',
}

export default DataType;